+++ zerver/tests/test_home.py
@@ -178,6 +178,7 @@ class HomeTest(ZulipTestCase):
         "realm_notifications_stream_id",
         "realm_password_auth_enabled",
         "realm_plan_type",
+        "realm_playgrounds",
         "realm_presence_disabled",
         "realm_private_message_policy",
         "realm_push_notifications_enabled",
@@ -262,7 +263,7 @@ class HomeTest(ZulipTestCase):
             set(result["Cache-Control"].split(", ")), {"must-revalidate", "no-store", "no-cache"}
         )
 
-        self.assert_length(queries, 40)
+        self.assert_length(queries, 41)
         self.assert_length(cache_mock.call_args_list, 5)
 
         html = result.content.decode("utf-8")
@@ -342,7 +343,7 @@ class HomeTest(ZulipTestCase):
                 result = self._get_home_page()
                 self.check_rendered_logged_in_app(result)
                 self.assert_length(cache_mock.call_args_list, 6)
-            self.assert_length(queries, 37)
+            self.assert_length(queries, 38)
 
     def test_num_queries_with_streams(self) -> None:
         main_user = self.example_user("hamlet")
@@ -373,7 +374,7 @@ class HomeTest(ZulipTestCase):
         with queries_captured() as queries2:
             result = self._get_home_page()
 
-        self.assert_length(queries2, 35)
+        self.assert_length(queries2, 36)
 
         # Do a sanity check that our new streams were in the payload.
         html = result.content.decode("utf-8")